requirements.txt
requests>=2.31
beautifulsoup4>=4.12
readability-lxml>=0.8.1
trafilatura>=1.9.0
tldextract>=5.1
python-dateutil>=2.9
langdetect>=1.0.9
rapidfuzz>=3.7.0
markdown2>=2.4.12
playwright>=1.47.0
lxml>=5.2
lxml_html_clean>=0.2.0
matplotlib>=3.8
folium>=0.17.0   # opzionale per mappa
